+++ CHANGELOG.md
@@ -15,6 +15,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Added weAreHere prop to UriLabel.
 
+- Add extra row "topRow" before some in Table component, which have "topRowKey".
+
 ## [0.29.0] - 2020-11-30
 
 - Hiding table head, when Table is empty.